// src/types/student.tsx
export interface Student {
    student_id: string;
    first_name: string;
    surname: string;
    nationality: string;
    national_id: string;
    place_of_birth: string;
    citizenship: string;
    permanent_address: string;
    email_address: string;
    phone_numbers: string;
    contact_address: string;
    permanent_home_address: string;
    date_of_birth: Date;
    marital_status: string;
    religion: string;
    title: string;
    sex: string;
    radio_frequency_id: string;
}